chore(react-bridge-adapter-sdk): refactor WalletProfile to EvmWalletProfile, prepare Accounts Profile View
