+++ packages/elasticbottle-react-bridge-adapter-sdk/src/components/ui/icons/WalletAdapterIcon.tsx
